commit changes of past 2 days, see changelog for details most important changes:  - added MSXPrinterPort  - PSG sounds louder
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@621 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/DACSound.cc
@@ -30,7 +30,7 @@ void DACSound::init()
 	setVolume(Mixer::MAX_VOLUME);
 	reset();
 	int bufSize = Mixer::instance()->registerSound(this);
-	buf = new short[bufSize];
+	buf = new int[bufSize];
 	lastChanged=Emutime(CLOCK,0);
 }
 
@@ -102,9 +102,9 @@ void DACSound::setSampleRate (int sampleRate)
 }
 
 
-short* DACSound::updateBuffer(int length)
+int* DACSound::updateBuffer(int length)
 {
-  short* buffer = buf;
+  int* buffer = buf;
   int nrsamples;
   short sample;
   while (length >0 ){
